Factor out ScriptChildProcess class from ScriptExecution
